+++ sdk/formrecognizer/Azure.AI.FormRecognizer/tests/Models/FormRecognizerModelFactoryTests.cs
@@ -57,7 +57,7 @@ namespace Azure.AI.FormRecognizer.Tests
             var formRecognizerError = new FormRecognizerError("", "");
 
             var modelId = "18910691-1896-0619-1896-091118961109";
-            var displayName = "My model";
+            var modelName = "My model";
             var properties = new CustomFormModelProperties();
             var status = CustomFormModelStatus.Ready;
             var trainingStartedOn = DateTimeOffset.Parse("1723-07-31T23:29:31Z");
@@ -66,10 +66,10 @@ namespace Azure.AI.FormRecognizer.Tests
             var trainingDocuments = new List<TrainingDocumentInfo>() { trainingDocumentInfo };
             var errors = new List<FormRecognizerError>() { formRecognizerError };
 
-            var customFormModel = FormRecognizerModelFactory.CustomFormModel(modelId, status, trainingStartedOn, trainingCompletedOn, submodels, trainingDocuments, errors, displayName, properties);
+            var customFormModel = FormRecognizerModelFactory.CustomFormModel(modelId, status, trainingStartedOn, trainingCompletedOn, submodels, trainingDocuments, errors, modelName, properties);
 
             Assert.AreEqual(modelId, customFormModel.ModelId);
-            Assert.AreEqual(displayName, customFormModel.DisplayName);
+            Assert.AreEqual(modelName, customFormModel.ModelName);
             Assert.AreEqual(status, customFormModel.Status);
             Assert.AreEqual(trainingStartedOn, customFormModel.TrainingStartedOn);
             Assert.AreEqual(trainingCompletedOn, customFormModel.TrainingCompletedOn);
@@ -100,16 +100,16 @@ namespace Azure.AI.FormRecognizer.Tests
         public void FormRecognizerModelFactoryCanInstantiateCustomFormModelInfo()
         {
             var modelId = "10001112-2233-3444-5556-667778889991";
-            var displayName = "My model";
+            var modelName = "My model";
             var properties = new CustomFormModelProperties();
             var trainingStartedOn = DateTimeOffset.Parse("1933-05-17T11:59:02Z");
             var trainingCompletedOn = DateTimeOffset.Parse("1949-03-19T07:41:47Z");
             var status = CustomFormModelStatus.Ready;
 
-            var customFormModelInfo = FormRecognizerModelFactory.CustomFormModelInfo(modelId, trainingStartedOn, trainingCompletedOn, status, displayName, properties);
+            var customFormModelInfo = FormRecognizerModelFactory.CustomFormModelInfo(modelId, trainingStartedOn, trainingCompletedOn, status, modelName, properties);
 
             Assert.AreEqual(modelId, customFormModelInfo.ModelId);
-            Assert.AreEqual(displayName, customFormModelInfo.DisplayName);
+            Assert.AreEqual(modelName, customFormModelInfo.ModelName);
             Assert.AreEqual(properties, customFormModelInfo.Properties);
             Assert.AreEqual(trainingStartedOn, customFormModelInfo.TrainingStartedOn);
             Assert.AreEqual(trainingCompletedOn, customFormModelInfo.TrainingCompletedOn);